Added the Welcome card for new users. It will be shown when user installs the application for the first time on his/her device.
