+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.stayfit.app.stayfit">
+    package="com.stayfit.app.stayfitBharat">
 
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
@@ -28,7 +28,7 @@
             android:value="${mapsKey}" />
 
         <activity
-            android:name=".MainActivity"
+            android:name="com.stayfit.app.stayfitBharat.MainActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -36,14 +36,14 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".WelcomeActivity"></activity>
-        <activity android:name=".SignUp" />
-        <activity android:name=".SignUpGoal" />
+        <activity android:name="com.stayfit.app.stayfitBharat.WelcomeActivity"></activity>
+        <activity android:name="com.stayfit.app.stayfitBharat.SignUp" />
+        <activity android:name="com.stayfit.app.stayfitBharat.SignUpGoal" />
         <activity
-            android:name=".FragmentActivity"
+            android:name="com.stayfit.app.stayfitBharat.FragmentActivity"
             android:label="@string/title_activity_fragment" />
         <activity
-            android:name=".MapsActivity"
+            android:name="com.stayfit.app.stayfitBharat.MapsActivity"
             android:label="@string/title_activity_maps" />
     </application>
