+++ app/src/main/java/com/stayfit/app/stayfitBharat/AddFoodToDiaryFragment.java
@@ -1,4 +1,4 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -21,8 +21,6 @@ import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
-
-import com.stayfit.app.stayfit.R;
 
 import java.util.ArrayList;
