+++ app/src/main/java/com/stayfit/app/stayfitBharat/DataParser.java
@@ -1,4 +1,4 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import org.json.JSONArray;
 import org.json.JSONException;
