+++ app/src/main/java/com/stayfit/app/stayfitBharat/DownloadUrl.java
@@ -1,4 +1,4 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import java.io.BufferedReader;
 import java.io.IOException;
