+++ app/src/main/java/com/stayfit/app/stayfitBharat/GoalFragment.java
@@ -1,4 +1,4 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -21,7 +21,6 @@ import android.widget.TableRow;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.stayfit.app.stayfit.R;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.text.DateFormat;
