+++ app/src/main/java/com/stayfit/app/stayfitBharat/ProfileFragment.java
@@ -1,4 +1,4 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -20,8 +20,6 @@ import android.widget.RadioGroup;
 import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
-
-import com.stayfit.app.stayfit.R;
 
 import java.util.Calendar;
