+++ app/src/main/java/com/stayfit/app/stayfitBharat/TestFragment.java
@@ -1,25 +1,24 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
+
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.stayfit.app.stayfit.R;
-
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link UpdateFragment.OnFragmentInteractionListener} interface
+ * {@link TestFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link UpdateFragment#newInstance} factory method to
+ * Use the {@link TestFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class UpdateFragment extends Fragment {
+public class TestFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -31,7 +30,7 @@ public class UpdateFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public UpdateFragment() {
+    public TestFragment() {
         // Required empty public constructor
     }
 
@@ -41,31 +40,17 @@ public class UpdateFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment UpdateFragment.
+     * @return A new instance of fragment TestFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static UpdateFragment newInstance(String param1, String param2) {
-        UpdateFragment fragment = new UpdateFragment();
+    public static TestFragment newInstance(String param1, String param2) {
+        TestFragment fragment = new TestFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
     }
-
-
-    /*- 05 on Activity Created ---------------------------------------------------------- */
-    // Run methods when started
-    // Set toolbar menu items
-    @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-
-        /* Set title */
-        ((FragmentActivity) getActivity()).getSupportActionBar().setTitle("Update");
-
-    } // onActivityCreated
-
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -80,7 +65,7 @@ public class UpdateFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_update, container, false);
+        return inflater.inflate(R.layout.fragment_test, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
