+++ app/src/main/java/com/stayfit/app/stayfitBharat/UpdateFragment.java
@@ -1,26 +1,23 @@
-package com.stayfit.app.stayfit;
+package com.stayfit.app.stayfitBharat;
 
 import android.content.Context;
 import android.net.Uri;
 import android.os.Bundle;
-
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.stayfit.app.stayfit.R;
-
 
 /**
  * A simple {@link Fragment} subclass.
  * Activities that contain this fragment must implement the
- * {@link TestFragment.OnFragmentInteractionListener} interface
+ * {@link UpdateFragment.OnFragmentInteractionListener} interface
  * to handle interaction events.
- * Use the {@link TestFragment#newInstance} factory method to
+ * Use the {@link UpdateFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class TestFragment extends Fragment {
+public class UpdateFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -32,7 +29,7 @@ public class TestFragment extends Fragment {
 
     private OnFragmentInteractionListener mListener;
 
-    public TestFragment() {
+    public UpdateFragment() {
         // Required empty public constructor
     }
 
@@ -42,17 +39,31 @@ public class TestFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment TestFragment.
+     * @return A new instance of fragment UpdateFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static TestFragment newInstance(String param1, String param2) {
-        TestFragment fragment = new TestFragment();
+    public static UpdateFragment newInstance(String param1, String param2) {
+        UpdateFragment fragment = new UpdateFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
         fragment.setArguments(args);
         return fragment;
     }
+
+
+    /*- 05 on Activity Created ---------------------------------------------------------- */
+    // Run methods when started
+    // Set toolbar menu items
+    @Override
+    public void onActivityCreated(Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
+
+        /* Set title */
+        ((FragmentActivity) getActivity()).getSupportActionBar().setTitle("Update");
+
+    } // onActivityCreated
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -67,7 +78,7 @@ public class TestFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_test, container, false);
+        return inflater.inflate(R.layout.fragment_update, container, false);
     }
 
     // TODO: Rename method, update argument and hook method into UI event
