+++ app/src/main/res/layout/activity_maps.xml
@@ -70,7 +70,7 @@ xmlns:android="http://schemas.android.com/apk/res/android">
     android:layout_alignParentBottom="true"
     android:layout_alignParentLeft="true"
     android:layout_alignParentStart="true"
-    tools:context="com.stayfit.app.stayfit.MapsActivity"
+    tools:context="com.stayfit.app.stayfitBharat.MapsActivity"
     android:layout_below="@+id/linearLayout2"/>
