+++ app/src/main/res/layout/content_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.stayfit.app.stayfit.FragmentActivity"
+    tools:context="com.stayfit.app.stayfitBharat.FragmentActivity"
     tools:showIn="@layout/app_bar_fragment">
 
 </android.support.constraint.ConstraintLayout>
