+++ app/src/main/res/layout/fragment_add_food_to_diary.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.stayfit.app.stayfit.CategoriesFragment">
+    tools:context="com.stayfit.app.stayfitBharat.CategoriesFragment">
 
     <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
