+++ app/src/main/res/layout/fragment_test.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.stayfit.app.stayfit.TestFragment">
+    tools:context="com.stayfit.app.stayfitBharat.TestFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
