+++ app/src/main/res/layout/fragment_update.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.stayfit.app.stayfit.UpdateFragment">
+    tools:context="com.stayfit.app.stayfitBharat.UpdateFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
